// Project-Portal/src/app/Models/accidents.ts
export class Accidents {
  id: number;
  source: string;
  latitude: string;
  longitude: string;
  description: string;
  street: string;
  city: string;
  state: string;
  county: string;
  zipcode: string;
  startTime: string;
  humidity: number;
  visibility: number;
  weathercondition: string;
}